Split pytest from source in matrix multiplier example
